+++ src/features/auth/SignUp/components/SignUpText.tsx
