+++ src/features/home/components/common/Banner.tsx
