+++ src/features/home/components/common/RecruitmentCard/index.tsx
@@ -2,8 +2,8 @@ import { RecruitmentItem } from '@/types';
 import { bounceAnimation } from '@assets/styles/animations';
 import { Card } from '@components/common';
 import { ReactNode } from 'react';
-import { RecruitmentCardContextProvider } from './RecruitmentCard.context';
-import { Title, Button, CompanyName, CompanyImage, Detail, Salary } from '.';
+import { RecruitmentCardContextProvider } from './index.context';
+import { Title, Button, CompanyName, CompanyImage, Detail, Salary } from './sub-components';
 import { responsiveStyle } from '@utils/responsive';
 
 const recruitmentCardStyle = responsiveStyle({
