+++ src/features/home/components/common/RecruitmentCard/sub-components/index.ts
