+++ src/features/home/components/common/RecruitmentHeader/index.styles.ts
