+++ src/pages/auth/SignIn/index.tsx
@@ -2,8 +2,8 @@ import { InnerContainer, Flex } from '@components/common';
 import Layout from '@features/layout';
 import Illustration from '@assets/images/signin-Illustration.svg?react';
 import { sectionStyle, innerContainerStyle, catchphraseContainerStyle } from './index.styles';
-import { SignInText } from '@/features/auth/SignIn/SignInText';
-import { SignInButton } from '@/features/auth/SignIn/SignInButton';
+import { SignInText } from '@/features/auth/SignIn/components/SignInText';
+import { SignInButton } from '@/features/auth/SignIn/components/SignInButton';
 
 export default function SignIn() {
   return (
