+++ src/types/select.ts
@@ -1,0 +1,5 @@
+export type SelectOptionType = {
+  value: string;
+  text: string;
+  action: () => void;
+};
